extract shared proxy machinery into lib/core
